Bug #25066: exec sessreg
Sessreg saves the pid of the calling process in utmp.
If this process is terminated the session is implicitly deleted.
Xstartup is called and the process id is gone right when
the script is done so the session is closed right away.
Using "exec" will save the pid of the master process of this login.

Signed-off-by: Johannes Dewender <xorg@JonnyJD.net>
Tested-by: Johannes Dewender <xorg@JonnyJD.net>
Acked-by: Julien Cristau <jcristau@debian.org>
Signed-off-by: Alan Coopersmith <alan.coopersmith@sun.com>