man: build man page using sed rather than the C preprocessor
This fixes misalignments to correctly display \n and \ for sample file.
No source code changes needed to fix that.

Replace XCOMM IMakefile comment with #.

Replace #if with:
.if !'x.VARNAME'x.' .ig
[man page text]
..

If $(VARNAME) is empty, the man page text is shown
If $(VARNAME) is not empty, the man page text is skipped

Or replace #if with:
.if 'VARNAME'' .ig
[man page text]
..

If $(VARNAME) is empty, the man page text is skipped
If $(VARNAME) is not empty, the man page text is shown

In DisplayManager.randomDevice, the default /dev/urandom is no longer
displayed with double-quotes which is consistent with other
default values. This was a cpp side-effect.

Tested on Debian with groff and Solaris with nroff

Co-authored-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>