config: use libtool -export-dynamic option for reverse dependencies
This replaces -rdynamic which is a GNU/Linux only solution.

"If symbols from your executable are needed to satisfy unresolved
references in a library you want to dlopen you will have to use
the flag -export-dynamic. You should use -export-dynamic while
linking the executable that calls dlopen."

It is used by the xserver, in Xdmx for example.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>