Stop leaking address struct on every call to getLocalAddress()
getLocalAddress() checks the haveLocalAddress static variable to see
if it's already set up an address struct, and if not allocates memory
to do so and stores a pointer in the localAddress static variable.

Unfortunately, it never then set haveLocalAddress, so every time it was
called it simply overwrote the previous allocation with a new one.

Now we check to see if the allocation succeeded, and if so, then set
haveLocalAddress so we don't repeat and waste time & memory on later
calls.

Reported-by: Ilja Van Sprundel <ivansprundel@ioactive.com>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>