Only pass -u & -w args to sessreg if --with-{u,w}tmp-file=path specified
If not specified, let sessreg use its builtin defaults instead of
replicating the logic here (and possibly getting it wrong), especially
since in sessreg-1.0.8 and earlier, using the -u & -w flags forces
the use of the utmp/wtmp code and not the newer utmpx/wtmpx code.

If --without-{u,w}tmp-file or --with-{u,w}tmp-file=none is specified,
pass "none" as arguments to sessreg to disable writing to that file.

v2: better handling of --with & --without arguments when not passing path,
add documentation of the 3 choices to README

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>