Only free input or output to realloc, not both
In error cleanup path, check to see if realloc returned NULL,
if so, free the still valid input pointer; otherwise the input is
no longer valid, so ignore it and free the realloc output pointer.
Avoids potential double free if you run out of memory at the exact
right spot.

Introduced-by: commit d39dbde390f0f609c03dbd79f0bcc6a3486ca8a3
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>